+++ src/scenes/Admissions/styles.css
@@ -1,8 +1,19 @@
 #admissions {
-  margin-top: 100px;
+  width: 100%;
+  margin-top: 60px;
 }
 
 #search {
   position: relative;
-  width: 100%;
+  width: 500px;
+  justify-content: flex-start;
+  border: solid 1px gray;
+  border-radius: 5px;
+  margin-bottom: 30px;
+}
+
+#search-icon {
+  height: 25px;
+  width: 25px;
+  margin-right: 10px;
 }
